fix: allow newer DNF versions to validate RPMs
Newer (dnf) versions seems to require both a V3 and V4 signature header.
The absence of a V3 header (or full payload digest tag) fails the
signature checks and so the RPM can be installed using plain `rpm`, but
not using `dnf install` from a repository.

This change adds support for V3 (header and payload) signatures, as well
as support for full payload digests (raw and compressed).

Closes: https://github.com/eclipse/packager/issues/42
Closes: https://github.com/eclipse/packager/issues/43