refactor: fix up second unit test, adding hash algorithm parameter
This change adds support for testing the newly created signature
functionality. It also adds the requirement to specify the hash
algorithm used, as the default that was used (SHA1) was no longer
supported by the rpm command.

The test uses podman by default for testing, as it is necessary to
import the GPG key for testing, and we don't want to mess with the
system running the tests.