Enhance type safety when getting tag values (#63)
This removes all uses of `ReadableHeader::getValue` which returns
`Object` and replaces them with methods that return specific types.
This simplifies the code and eliminates the need for the caller to
cast the result.

The method `RpmBaseTag::getDataType` was added to store the tag's
data type. The tag's data type is checked when the various
`ReadableHeader` methods are called.